Alteracao de campos smallint de int para short

+++ Models/Ordem.cs
@@ -4,9 +4,7 @@ namespace logiWeb.Models
     {
         public int Id { get; set; }
         public int IdCliente { get; set; }
-
-        public int Qtd_itens { get; set; }
-        
+        public short Qtd_itens { get; set; }
         public Funcionario Funcionario {get; set;} = new Funcionario();
         public Cliente Cliente {get; set;} = new Cliente();
         public string Destino { get; set; } = string.Empty;
